profile view, question request updates
